Restructuring, `commands.py`
